added: views in apps; fixed: urls
