+++ pay2u/api/urls.py
@@ -1,27 +1,30 @@
 from django.urls import path
 
-from .views import (
+from payments.view import (
     AccountPaymentView,
-    AccountView,
-    ActiveUserSubscriptionView,
-    AvailableServicesView,
-    СategoriesView,
-    CSRFTokenView,
     DocumentView,
+    PaymentView,
+    PaymentsPeriodView,
+    PaymentsView,
+    ServicePaymentsView
+)
+from subscriptions.view import (
+    ActiveUserSubscriptionView,
     MainPageView,
     NonActiveUserSubscriptionView,
-    PaymentsPeriodView,
-    PaymentView,
-    PaymentsView,
-    ServicePaymentsView,
-    ServiceView,
-    UserPaymentsPlanView,
     ServiceUserSubscriptionsView,
+    UserPaymentsPlanView,
     UserSubscriptionRenewalView,
-    ActiveUserSubscriptionView,
-    AddUserSubscriptionView,
     UserSubscriptionView,
-    UserSubscriptionsView,
+    UserSubscriptionsView
+)
+from users.view import AccountView
+from .views import (
+    AvailableServicesView,
+    CategoriesView,
+    CSRFTokenView,
+    ServiceView,
+    AddUserSubscriptionView
 )
 
 urlpatterns = [
@@ -105,7 +108,7 @@ urlpatterns = [
     ),
     path(
         "v1/categories/<str:category_name>/",
-        СategoriesView.as_view(),
+        CategoriesView.as_view(),
         name="categories",
     ),
     path(
